backend/requirements.txt
bcrypt==4.0.1
email-validator==2.1.0.post1
fastapi==0.95.1
httpx==0.24.1
Jinja2==3.1.2
pydantic==1.10.13
pyodbc==5.0.1
pytest==7.4.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
SQLAlchemy==2.0.13
uvicorn==0.22.0
alembic==1.11.1
passlib==1.7.4
